+++ hoomd-microstate/src/boundary/periodic/eighteight.rs
@@ -86,7 +86,7 @@ impl Wrap<Point<Hyperbolic<3>>> for Periodic<EightEight> {
     /// # Ok(())
     /// # }
     /// ```
-    #[inline]
+    #[inline(always)]
     #[expect(clippy::too_many_lines, reason = "complicated function")]
     fn wrap(&self, properties: Point<Hyperbolic<3>>) -> Result<Point<Hyperbolic<3>>, Error> {
         let mut properties = properties;
@@ -422,7 +422,7 @@ impl GenerateGhosts<Point<Hyperbolic<3>>> for Periodic<EightEight> {
         self.maximum_interaction_range
     }
     /// Place periodic images of sites near the edge of the periodic boundary
-    #[inline]
+    #[inline(always)]
     fn generate_ghosts(
         &self,
         site_properties: &Point<Hyperbolic<3>>,
@@ -495,7 +495,7 @@ impl GenerateGhosts<OrientedHyperbolicPoint<3, Angle>> for Periodic<EightEight> 
         self.maximum_interaction_range
     }
     /// Place periodic images of sites near the edge of the periodic boundary
-    #[inline]
+    #[inline(always)]
     #[expect(clippy::too_many_lines, reason = "complicated function")]
     fn generate_ghosts(
         &self,