+++ hoomd-microstate/src/boundary/periodic/triclinic.rs
@@ -96,7 +96,7 @@ where
     /// # Ok(())
     /// # }
     /// ```
-    #[inline]
+    #[inline(always)]
     fn wrap(&self, mut properties: P) -> Result<P, Error> {
         let r = properties.position_mut();
         let mut fractional = self.shape.fractional(r);
@@ -173,7 +173,7 @@ where
         clippy::too_many_lines,
         reason = "There are many (literal) corner cases."
     )]
-    #[inline]
+    #[inline(always)]
     fn generate_ghosts(&self, site_properties: &S) -> ArrayVec<S, MAX_GHOSTS> {
         let mut result = ArrayVec::new();
         let r: &Cartesian<3> = site_properties.position();